maxPinnedFonts should be int instead of float.

+++ rts/Rendering/Fonts/CFontTexture.h
@@ -187,7 +187,7 @@ private:
 	int lastTextureUpdate = 0;
 	bool needsTextureUpload = true;
 	inline static int maxFontTries = 0;
-	inline static float maxPinnedFonts = 0;
+	inline static int maxPinnedFonts = 0;
 #endif
 	std::shared_ptr<FontFace> shFace;
